requirements.txt
# Web framework and server
fastapi==0.118.0
uvicorn==0.37.0
starlette==0.48.0
anyio==4.11.0
httptools==0.6.4
watchfiles==1.1.0
websockets==15.0.0

# Database and ORM
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
greenlet==3.2.4

# API and Request handling
python-multipart==0.0.20
httpx==0.28.1
requests==2.32.5
urllib3==2.3.0

# Data validation and settings
pydantic==2.11.10
pydantic-core==2.33.2
python-dotenv==1.1.1
annotated-types==0.7.0

# AI and Google Generative AI
google-cloud-aiplatform==1.119.0
google-generativeai==0.8.5
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-storage==2.19.0
langchain==0.3.27
langchain-core==0.3.78
langsmith==0.4.32

# Vector store and embeddings
chromadb==1.1.0
tiktoken==0.11.0
tokenizers==0.22.1
huggingface-hub==0.35.3

# Document processing
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
beautifulsoup4==4.14.2
lxml==6.0.2

# Testing
pytest==8.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities and helpers
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.15.0
coloredlogs==15.0.1
rich==14.1.0
orjson==3.11.3
numpy==2.3.3
pybase64==1.4.2